+++ js/dashboard.js
@@ -25,7 +25,7 @@ OCA.Analytics = Object.assign({}, OCA.Analytics, {
     TYPE_GIT: 3,
     TYPE_EXTERNAL_FILE: 4,
     TYPE_EXTERNAL_REGEX: 5,
-    TYPE_EXCEL: 7,
+    TYPE_SPREADSHEET: 7,
     TYPE_SHARED: 99,
     // flexible mapping depending on type required by the used chart library
     chartTypeMapping: {