+++ js/panorama.js
@@ -41,7 +41,7 @@ OCA.Analytics = Object.assign({}, OCA.Analytics, {
     TYPE_GIT: 3,
     TYPE_EXTERNAL_FILE: 4,
     TYPE_EXTERNAL_REGEX: 5,
-    TYPE_EXCEL: 7,
+    TYPE_SPREADSHEET: 7,
     TYPE_SHARED: 99,
     SHARE_TYPE_USER: 0,
     SHARE_TYPE_GROUP: 1,