+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
 ### Changed
 - PhpSpreadsheet 3.4.0
+- local spreadsheet: handle ranges
 
 ### Fixed
 - wrong info link in data load screen